\name{PMA-internal}
\title{Internal PMA functions}
\alias{print.PMDL1L1}
\alias{print.PMDL1L1.cv}
\alias{print.PMDL1FL}
\alias{print.PMDL1FL.cv}
\alias{plot.PMDL1L1.cv}
\alias{plot.PMDL1FL.cv}
\description{Internal PMA functions}
\usage{
\method{print}{PMDL1L1}(x,verbose=FALSE,...)
\method{print}{PMDL1L1.cv}(x,...)
\method{plot}{PMDL1L1.cv}(x,...)
\method{print}{PMDL1FL}(x,verbose=FALSE,...)
\method{print}{PMDL1FL.cv}(x,...)
\method{plot}{PMDL1FL.cv}(x,...)
}
\author{Daniela M. Witten and Robert Tibshirani}
\keyword{internal}




