\name{PMA-internal}
\title{Internal PMA functions}
\alias{print.spccv}
\alias{plot.spccv}
\alias{print.spc}
\alias{plot.ccaperml1l1}
\alias{plot.ccaperml1fl}
\alias{print.pmdl1l1}
\alias{print.ccaperml1fl}
\alias{print.pmdl1l1cv}
\alias{print.pmdl1fl}
\alias{print.pmdl1flcv}
\alias{print.ccal1l1}
\alias{print.ccaperml1l1}
\alias{plot.pmdl1l1cv}
\alias{plot.pmdl1flcv}
\alias{print.ccal1fl}
\description{Internal PMA functions}
\usage{
plot.spccv(x,...)
print.spccv(x,...)
print.spc(x,verbose=FALSE,...)
plot.ccaperml1l1(x,...)
plot.ccaperml1fl(x,...)
print.pmdl1l1(x,verbose=FALSE,...)
print.pmdl1l1cv(x,...)
plot.pmdl1l1cv(x,...)
plot.pmdl1flcv(x,...)
print.pmdl1fl(x,verbose=FALSE,...)
print.pmdl1flcv(x,...)
print.ccal1l1(x,verbose=FALSE,...)
print.ccaperml1l1(x,...)
print.ccaperml1fl(x,...)
print.ccal1fl(x,verbose=FALSE,...)
}
\author{Daniela M. Witten and Robert Tibshirani}
\keyword{internal}




