\name{PMA-internal}
\title{Internal PMA functions}
\alias{print.PMDL1L1}
\alias{print.PMDL1L1.cv}
\alias{print.PMDL1FL}
\alias{print.PMDL1FL.cv}
\alias{plot.PMDL1L1.cv}
\alias{plot.PMDL1FL.cv}
\alias{cor.xl.LoadData}
\alias{cor.xl.RunAnalysis}
\alias{cor.xl.Plot1}
\alias{cor.xl.Plot2}
\alias{cor.xl.Plot3}
\alias{cor.xl.Plot2a}
\alias{cor.xl.Plot2b}
\alias{cor.xl.Plot3a}
\alias{cor.xl.Plot3b}
\description{Internal PMA functions}
\usage{
\method{print}{PMDL1L1}(x,verbose=FALSE,...)
\method{print}{PMDL1L1.cv}(x,...)
\method{plot}{PMDL1L1.cv}(x,...)
\method{print}{PMDL1FL}(x,verbose=FALSE,...)
\method{print}{PMDL1FL.cv}(x,...)
\method{plot}{PMDL1FL.cv}(x,...)
cor.xl.LoadData(bl.col, smp.row, which)
cor.xl.RunAnalysis(boollst1, boollst2, K)
cor.xl.Plot1()
cor.xl.Plot2()
cor.xl.Plot3()
cor.xl.Plot2a()
cor.xl.Plot2b()
cor.xl.Plot3a()
cor.xl.Plot3b()
}
\author{Daniela M. Witten and Robert Tibshirani}
\keyword{internal}




