\name{PMA-internal}
\title{Internal PMA functions}
\alias{print.multicca}
\alias{print.multiccaperm}
\alias{plot.multiccaperm}
\alias{print.spccv}
\alias{plot.spccv}
\alias{print.spc}
\alias{plot.ccaperm}
\alias{print.pmdl1l1}
\alias{print.ccaperm}
\alias{print.pmdl1l1cv}
\alias{print.pmdl1fl}
\alias{print.pmdl1flcv}
\alias{print.cca}
\alias{plot.pmdl1l1cv}
\alias{plot.pmdl1flcv}
\alias{cor.xl.LoadData}
\alias{cor.xl.RunAnalysis}
\alias{cor.xl.Plot1}
\alias{cor.xl.Plot2}
\alias{cor.xl.Plot3}
\alias{cor.xl.Plot2a}
\alias{cor.xl.Plot2b}
\alias{cor.xl.Plot3a}
\alias{cor.xl.Plot3b}
\description{Internal PMA functions}
\usage{
plot.multiccaperm(x,...)
print.multicca(x,...)
print.multiccaperm(x,...)
plot.spccv(x,...)
print.spccv(x,...)
print.spc(x,verbose=FALSE,...)
plot.ccaperm(x,...)
print.pmdl1l1(x,verbose=FALSE,...)
print.pmdl1l1cv(x,...)
plot.pmdl1l1cv(x,...)
plot.pmdl1flcv(x,...)
print.pmdl1fl(x,verbose=FALSE,...)
print.pmdl1flcv(x,...)
print.cca(x,verbose=FALSE,...)
print.ccaperm(x,...)
cor.xl.LoadData(bl.col, smp.row, which)
cor.xl.RunAnalysis(boollst1, boollst2, K)
cor.xl.Plot1()
cor.xl.Plot2()
cor.xl.Plot3()
cor.xl.Plot2a()
cor.xl.Plot2b()
cor.xl.Plot3a()
cor.xl.Plot3b()
}
\author{Daniela M. Witten and Robert Tibshirani}
\keyword{internal}




